% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod-month-birthstone.R
\name{birthstone_ui}
\alias{birthstone_ui}
\title{UI for birthstone}
\usage{
birthstone_ui(id)
}
\arguments{
\item{id}{UI id}
}
\value{
html and text string with birthstone summary
}
\description{
UI for birthstone
}
