% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/month_mod_birthstone.R
\name{birthstoneServer}
\alias{birthstoneServer}
\title{server function for birthstone}
\usage{
birthstoneServer(id, month, stone_df)
}
\arguments{
\item{id}{server id}

\item{month}{a vector of month names}

\item{stone_df}{a data frame with a column for month and a column for stone}
}
\value{
outputs for month and stone based on input
}
\description{
server function for birthstone
}
