% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/month_mod_birthstone.R
\name{birthstoneUI}
\alias{birthstoneUI}
\title{UI for birthstone}
\usage{
birthstoneUI(id)
}
\arguments{
\item{id}{UI id}
}
\value{
html and text string with birthstone summary
}
\description{
UI for birthstone
}
