% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app-month.R
\name{month_app}
\alias{month_app}
\title{toy app from mastering shiny, chap 20}
\usage{
month_app(...)
}
\arguments{
\item{...}{options pass to \code{shinyApp}.}
}
\description{
toy app from mastering shiny, chap 20
}
