% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod-month_birthstone.R
\name{month_birthstone_server}
\alias{month_birthstone_server}
\title{server function for birthstone}
\usage{
month_birthstone_server(id, month, stone_df)
}
\arguments{
\item{id}{server id}

\item{month}{a vector of month names}

\item{stone_df}{a data frame with a column for month and a column for stone}
}
\value{
outputs for month and stone based on input
}
\description{
server function for birthstone
}
