% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod-month_feedback.R
\name{month_feedback_server}
\alias{month_feedback_server}
\title{serve for month}
\usage{
month_feedback_server(id, month)
}
\arguments{
\item{id}{ID}

\item{month}{unit character vector with chosen month}
}
\value{
string with month feedback
}
\description{
serve for month
}
