% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/month_mod_feedback.R
\name{monthFeedbackServer}
\alias{monthFeedbackServer}
\title{serve for month}
\usage{
monthFeedbackServer(id, month)
}
\arguments{
\item{id}{ID}

\item{month}{unit character vector with chosen month}
}
\value{
string with month feedback
}
\description{
serve for month
}
