% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app_data.R
\name{data_app}
\alias{data_app}
\title{toy app from mastering shiny, chap 1}
\usage{
data_app(...)
}
\arguments{
\item{...}{Options passes to \code{shinyApp}.}
}
\description{
toy app from mastering shiny, chap 1
}
